% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadunico.R
\name{available_cadunico}
\alias{available_cadunico}
\title{List available CadUnico datasets}
\usage{
available_cadunico()
}
\value{
A tibble with columns: year, description, source, reference_date.
}
\description{
Returns a tibble with all available CadUnico sample datasets.
}
\examples{
available_cadunico()
}
