% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_suas.R
\name{available_censo_suas}
\alias{available_censo_suas}
\title{List available Censo SUAS datasets}
\usage{
available_censo_suas()
}
\value{
A tibble with columns: year, questionnaire, description, source.
}
\description{
Returns a tibble with all available Censo SUAS datasets, including
years and questionnaire types.
}
\examples{
available_censo_suas()
}
