% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadunico.R
\name{build_cadunico_url}
\alias{build_cadunico_url}
\title{Build CadUnico download URL}
\usage{
build_cadunico_url(year)
}
\arguments{
\item{year}{Integer. Year.}
}
\value{
Character. Download URL.
}
\description{
Build CadUnico download URL
}
\keyword{internal}
