% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_suas.R
\name{build_censo_suas_url}
\alias{build_censo_suas_url}
\title{Build Censo SUAS download URL}
\usage{
build_censo_suas_url(year, questionnaire)
}
\arguments{
\item{year}{Integer. Year.}

\item{questionnaire}{Character. Questionnaire type (not used for 2011-2017,
as all questionnaires are in a single archive).}
}
\value{
Character. Download URL.
}
\description{
Build Censo SUAS download URL
}
\keyword{internal}
