% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_url}
\alias{build_url}
\title{Build URL from base and parameters}
\usage{
build_url(base_url, ...)
}
\arguments{
\item{base_url}{Character. Base URL.}

\item{...}{Named parameters to append to URL.}
}
\value{
Character. Complete URL.
}
\description{
Build URL from base and parameters
}
\keyword{internal}
