% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{check_cache}
\alias{check_cache}
\title{Check if data exists in cache}
\usage{
check_cache(dataset, ...)
}
\arguments{
\item{dataset}{Character. Dataset name.}

\item{...}{Additional identifiers (year, questionnaire, etc.).}
}
\value{
Character or NULL. Path to cached file if exists, NULL otherwise.
}
\description{
Check if data exists in cache
}
\keyword{internal}
