% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadunico.R
\name{create_cadunico_survey}
\alias{create_cadunico_survey}
\title{Create survey design object for CadUnico}
\usage{
create_cadunico_survey(data)
}
\arguments{
\item{data}{A tibble with CadUnico data.}
}
\value{
A srvyr survey design object.
}
\description{
Create survey design object for CadUnico
}
\keyword{internal}
