% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadunico.R
\name{dictionary_cadunico}
\alias{dictionary_cadunico}
\title{Get CadUnico data dictionary}
\usage{
dictionary_cadunico(year)
}
\arguments{
\item{year}{Integer. Year of the sample.}
}
\value{
A tibble with columns: variable, description, type, values.
}
\description{
Returns the data dictionary (variable descriptions) for a specific
CadUnico sample dataset.
}
\details{
The official data dictionary is available at the Brazilian Open Data Portal:
\url{https://dados.gov.br/dados/conjuntos-dados/microdados-amostrais-do-cadastro-unico}
}
\examples{
\dontrun{
# get dictionary for cadunico 2018
dict <- dictionary_cadunico(2018)
}
}
