% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_chunked}
\alias{download_chunked}
\title{Download large files in chunks}
\usage{
download_chunked(urls, destdir = NULL, quiet = FALSE)
}
\arguments{
\item{urls}{Character vector. URLs to download.}

\item{destdir}{Character. Destination directory.}

\item{quiet}{Logical. Suppress progress messages? Default FALSE.}
}
\value{
Character vector. Paths to downloaded files.
}
\description{
Download large files in chunks
}
\keyword{internal}
