% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_suas.R
\name{filter_by_uf}
\alias{filter_by_uf}
\title{Filter data by UF (state)}
\usage{
filter_by_uf(data, uf)
}
\arguments{
\item{data}{A tibble.}

\item{uf}{Character. UF code.}
}
\value{
Filtered tibble.
}
\description{
Filter data by UF (state)
}
\keyword{internal}
