% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get cache directory path}
\usage{
get_cache_dir()
}
\value{
Character. Path to cache directory.
}
\description{
Returns the current cache directory used by welfarebR. By default, uses
the user's cache directory as determined by \code{\link[rappdirs]{user_cache_dir}}.
}
\examples{
get_cache_dir()
}
