% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{load_from_cache}
\alias{load_from_cache}
\title{Load data from cache}
\usage{
load_from_cache(path)
}
\arguments{
\item{path}{Character. Path to cached file.}
}
\value{
Cached data.
}
\description{
Load data from cache
}
\keyword{internal}
