% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadunico.R
\name{read_cadunico_file}
\alias{read_cadunico_file}
\title{Read and process CadUnico file}
\usage{
read_cadunico_file(path, year)
}
\arguments{
\item{path}{Character. Path to downloaded file.}

\item{year}{Integer. Year of data.}
}
\value{
A tibble with processed data.
}
\description{
Read and process CadUnico file
}
\keyword{internal}
