% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo_suas.R
\name{read_censo_suas_file}
\alias{read_censo_suas_file}
\title{Read and process Censo SUAS file}
\usage{
read_censo_suas_file(path, year, questionnaire)
}
\arguments{
\item{path}{Character. Path to downloaded file.}

\item{year}{Integer. Year of data.}

\item{questionnaire}{Character. Questionnaire type.}
}
\value{
A tibble with processed data.
}
\description{
Read and process Censo SUAS file
}
\keyword{internal}
