% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save data to cache}
\usage{
save_to_cache(data, dataset, ...)
}
\arguments{
\item{data}{Data to cache.}

\item{dataset}{Character. Dataset name.}

\item{...}{Additional identifiers (year, questionnaire, etc.).}
}
\value{
Invisible. Path to cached file.
}
\description{
Save data to cache
}
\keyword{internal}
