% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate common inputs}
\usage{
validate_inputs(
  year = NULL,
  questionnaire = NULL,
  valid_years = NULL,
  valid_questionnaires = NULL
)
}
\arguments{
\item{year}{Integer. Year to validate.}

\item{questionnaire}{Character. Questionnaire type.}

\item{valid_years}{Integer vector. Valid years.}

\item{valid_questionnaires}{Character vector. Valid questionnaire types.}
}
\value{
Invisible TRUE if valid, otherwise throws error.
}
\description{
Validate common inputs
}
\keyword{internal}
