% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welfarebr-package.R
\docType{package}
\name{welfarebR-package}
\alias{welfarebR}
\alias{welfarebR-package}
\title{welfarebR: Access Brazilian Social Assistance Data}
\description{
Provides functions to download and process Brazilian social assistance data in tidy format. Includes data from Censo SUAS (CRAS, CREAS, Municipal Management) and CadUnico (deidentified sample). Data sources are from the Ministry of Social Development (MDS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sidneybissoli/welfarebR}
  \item Report bugs at \url{https://github.com/sidneybissoli/welfarebR/issues}
}

}
\author{
\strong{Maintainer}: Sidney Bissoli \email{sbissoli76@gmail.com}

}
\keyword{internal}
